Integrate image upload into products form create
