+++ src/app/components/imagePreview/ImagePreviewComponent.js
@@ -7,7 +7,7 @@ const ImagePreviewComponent = ({ image }) =>
         {image.name}
       </div>
       <div className="ImagePreviewComponent-content">
-        <img src={image.thumbnail} alt="" />
+        <div className="ImagePreviewComponent-content-image" style={{ backgroundImage: `url(${image.thumbnail})`, opacity: image.percent === 100 ? 1 : 0.5 }} />
       </div>
       <div className="ImagePreviewComponent-footer">
         {image.percent}
