+++ src/app/containers/ProductFormContainer.js
@@ -4,6 +4,8 @@ import productFormContainerHoc from 'hoc/productFormContainerHoc';
 import ProductFormComponent from 'components/productForm/ProductFormComponent';
 import ProductAboutComponent from 'components/productAbout/ProductAboutComponent';
 import NotificationComponent from 'components/notification/NotificationComponent';
+import ImagePreviewComponent from 'components/imagePreview/ImagePreviewComponent';
+import ImageUploadComponent from 'components/imageUpload/ImageUploadComponent';
 import { TitleIconComponent } from 'components/icons/IconsComponent';
 
 export class ProductFormContainer extends React.Component {
@@ -27,6 +29,18 @@ export class ProductFormContainer extends React.Component {
   }
 
   render() {
+    const config = {
+      signingUrl: '/s3/sign',
+      accept: 'image/*',
+      onUploadStart: this.props.onUploadStart,
+      onUploadProgress: this.props.onUploadProgress,
+      onUploadError: this.props.onUploadError,
+      onUploadFinish: this.props.onUploadFinish,
+      uploadRequestHeaders: { 'x-amz-acl': 'public-read' },
+      multiple: 'true',
+      server: 'http://localhost:8080/api',
+    };
+
     return (
       <div className="ProductFormContainerBlock">
         <div className="ProductFormContainerBlock-title">
@@ -42,13 +56,26 @@ export class ProductFormContainer extends React.Component {
                 <NotificationComponent isVisible message={notificationsFailure.message} status={notificationsFailure.status} />
               )}
             </div>
+
             <div className="ProductFormContainer-block-left">
               <ProductFormComponent
                 createProduct={this.createProduct}
                 categories={this.props.mappedCategories}
                 subCategories={this.props.mappedSubCategories}
-              />
+              >
+                {!this.props.images.data.length ? <div className="ProductFormContainer-imagesEmpty">No uploaded images</div> : null}
+                {this.props.images.data.length ? <label htmlFor="">Uploaded images</label> : null}
+                {this.props.images.data.length ?
+                  <div className="ProductFormContainer-images">
+                    {this.props.images.data.map((image, key) => <div key={key}><ImagePreviewComponent image={image} /></div>)}
+                  </div>
+                : null}
+                <div className="ProductFormContainer-upload">
+                  <ImageUploadComponent {...config} />
+                </div>
+              </ProductFormComponent>
             </div>
+
             <div className="ProductFormContainer-block-right">
               <ProductAboutComponent />
             </div>
@@ -60,6 +87,12 @@ export class ProductFormContainer extends React.Component {
 }
 
 ProductFormContainer.propTypes = {
+  images: PropTypes.array.isRequired,
+  onUploadStart: PropTypes.func.isRequired,
+  onUploadProgress: PropTypes.func.isRequired,
+  onUploadError: PropTypes.func.isRequired,
+  onUploadFinish: PropTypes.func.isRequired,
+
   notificationsSuccess: PropTypes.array.isRequired,
   notificationsFailure: PropTypes.array.isRequired,
   mappedCategories: PropTypes.object.isRequired,
