+++ src/app/hoc/productFormContainerHoc.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { formValueSelector } from 'redux-form';
 import * as actions from 'actions/index';
-import { transformCategories, getFailedNotifications, getSucceededNotifications, filterSubCategories } from 'utils/index';
+import { transformCategories, getFailedNotifications, getSucceededNotifications, filterSubCategories, generateFilePreviewAsync } from 'utils/index';
 
 const selector = formValueSelector('ProductFormComponent');
 
@@ -10,6 +10,11 @@ const mapStateToProps = (state) => ({
   productSubCategories: state.productSubCategoriesReducer.productSubCategories,
   formCategoryId: selector(state, 'categoryId'),
   productsActions: state.productsReducer.actions,
+
+  /**
+   * Image upload
+   */
+  images: state.imageUploadReducer.images,
 });
 
 const mapDispatchToProps = {
@@ -18,6 +23,14 @@ const mapDispatchToProps = {
   getProductSubCategoriesRequest: actions.getProductSubCategoriesRequest,
   getProductCategoriesIdle: actions.getProductCategoriesIdle,
   getProductSubCategoriesIdle: actions.getProductSubCategoriesIdle,
+
+  /**
+   * Image upload
+   */
+  startImageUpload: actions.startImageUpload,
+  progressImageUpload: actions.progressImageUpload,
+  errorImageUpload: actions.errorImageUpload,
+  finishImageUpload: actions.finishImageUpload,
 };
 
 const mergeProps = (stateProps, dispatchProps, ownProps) => {
@@ -30,11 +43,37 @@ const mergeProps = (stateProps, dispatchProps, ownProps) => {
   const notificationsFailure = getFailedNotifications([
     stateProps.productsActions.postProduct,
   ]);
+
+  /**
+   * Image upload
+   */
+  const onUploadStart = (file, next) => {
+    generateFilePreviewAsync(file)
+    .then((thumbnail) => dispatchProps.startImageUpload({ name: file.name, size: file.size, percent: 0, thumbnail }))
+    .then(() => next(file));
+  };
+  const onUploadProgress = (percent, message, file) => {
+    dispatchProps.progressImageUpload({ name: file.name, percent });
+  };
+  const onUploadError = (errorMessage, file) => {
+    dispatchProps.errorImageUpload({ name: file.name, errorMessage });
+  };
+  const onUploadFinish = ({ publicUrl }, file) => {
+    dispatchProps.finishImageUpload({ name: file.name, publicUrl });
+  };
   return Object.assign({
     notificationsSuccess,
     notificationsFailure,
     mappedCategories,
     mappedSubCategories,
+
+    /**
+     * Image upload
+     */
+    onUploadStart,
+    onUploadProgress,
+    onUploadError,
+    onUploadFinish,
   }, ownProps, stateProps, dispatchProps);
 };
