+++ src/assets/scss/components/_ImagePreviewComponent.scss
@@ -1,0 +1,25 @@
+.ImagePreviewComponentBlock {
+}
+
+.ImagePreviewComponent {
+  padding: $padding-base-horizontal;
+  border: 1px solid $c-bodyBackground;
+
+  &-title {
+    margin-bottom: $padding-base-horizontal;
+  }
+
+  &-content {
+    margin-bottom: $padding-base-horizontal;
+    &-image {
+      display: block;
+      max-width: 300px;
+      height: 120px;
+      background-size: cover;
+      background-position: center center;
+    }
+  }
+
+  &-footer {
+  }
+}
