+++ src/assets/scss/components/_ImageUploadComponent.scss
@@ -1,0 +1,8 @@
+.ImageUploadComponentBlock {
+}
+
+.ImageUploadComponent {
+  input {
+    background: $c-bodyBackground;
+  }
+}
