+++ src/assets/scss/containers/_ProductFormContainer.scss
@@ -31,4 +31,24 @@
       }
     }
   }
+
+  &-imagesEmpty {
+    background: $c-bodyBackground;
+    padding: $padding-base-horizontal;
+    margin-bottom: $padding-large-horizontal;
+  }
+
+  &-images {
+    @extend .row;
+    margin-bottom: $padding-large-horizontal;
+
+    & > div {
+      @extend .col-xs-4;
+      margin-bottom: $padding-large-horizontal;
+    }
+  }
+
+  &-upload {
+    margin-bottom: $padding-large-horizontal;
+  }
 }
